requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-countries==7.1
django-debug-toolbar==3.2.1
django-js-asset==1.2.2
django-modeltranslation==0.17.1
django-mptt==0.12.0
django-rosetta==0.9.6
django-widget-tweaks==1.4.8
idna==2.10
importlib-metadata==4.0.1
kombu==5.0.2
oauthlib==3.1.0
Pillow==8.2.0
pkg-resources==0.0.0
polib==1.1.1
prompt-toolkit==3.0.18
pycparser==2.20
PyJWT==2.0.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
